requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
dlib==19.24.2
mediapipe==0.10.8
scikit-learn==1.3.2
pillow==10.1.0
streamlit==1.29.0
joblib==1.3.2